The indicator is country specific.
Sorry I missed the second part of your email – the cancer screening variables can come in two formats:

1.	Have you ever done a mammogram/pap smear?
2.	When was the last time you got a pap smear? – sometimes this question is open-ended, i.e. the woman tells the enumerator how many years ago the last test was done; or, there are categories of years like <1 year, 1-2 years, 3 years, etc.

Sometimes 1. Is a filter question for 2. 

Let’s actually construct three variables – one that gives the share of women with the respective test, one with the age group the question was asked for (a string with the upper and lower bound like “15-59”) and another one with the reference period (a string taking on “ever” if only 1. Is available, “3” if you can construct a 3-year reference period from 2. and any other reference period if you cannot construct the 3 year reference period).

For the previously coded DHS pls use the existing code in the old dofiles to build this.